Moved DOCKER-ISOLATION iptables rule at the very end.
Instead of inserting the RETURN rule for the DOCKER-ISOLATION chain at
the beginning of the Agent module process, we now insert it after all
configuration has been received from Master. This ensures that the
RETURN rule is the last rule be executed, placing it at the top of the
DOCKER-ISOLATION chain.