Converted all `EXIT` to `LOG`.
While creating the `MasterManagerProcess` the master module would
`EXIT` on an error. This is detrimental to running GMOCK/GTEST since
GMOCK would exist on a failed test case that wasn't able to initialize
the Master module. Accordingly converted all `EXIT` to `LOG(ERROR)`
and returned a `nullptr` whenever we see an exception.

Removed `EXIT` from Agent manager and replaced with `LOG(ERROR)`.
The Agent module would `EXIT` if it encounters an error during module
creation. However, this was causing problems in terms of running GMOCK
since the tests would `EXIT` on a failed case of module creation.
Accordingly replaced the `EXIT` with a `LOG(ERROR)`.