Fixing overlay network configuration failure.
Currently, when an overlay network configuration fails the agent
reports the error to the master and doesn't retry. The assumption was
that that operator will have to intervene in such a scenario to fix
the problem. However, some of these errors can be transient. For
example, Docker not being up or the iptables being inaccessible (being
used by a different process, say Docker). In such cases it is
imperative to retry the configuration instead of just giving up.

Accordingly, when the agent sees a failure to configure an overlay
network it simply drops the register acknowledgment from the master.
This would force the agent to re-register with the master causing it
to try re-configuring the overlay network that failed in the previous
attempt.